Fix prefs

+++ trunk/webEdition/we/include/we_modules/users/we_users_unlock.inc.php
@@ -25,8 +25,9 @@ we_html_tools::protect();
 
 
 // prepare the queries, 4 as maximum.
-$_ids =  we_base_request::_(we_base_request::INTLISTA, 'we_cmd', '', 1); // we_cmd[1] is commaseperated list of ids
-$_tables = explode(",", $_REQUEST['we_cmd'][3]); // we_cmd[3] is commaseparated list of tables
+$_ids = we_base_request::_(we_base_request::INTLISTA, 'we_cmd', '', 1); // we_cmd[1] is commaseperated list of ids
+//FIXME: make tblList???
+$_tables = explode(',', we_base_request::_(we_base_request::STRINGC, 'we_cmd', '', 3)); // we_cmd[3] is commaseparated list of tables
 $_transaction = we_base_request::_(we_base_request::TRANSACTION, 'we_cmd', null, 4); // we_cmd[4] is a single transaction, to delete data from session
 
 $queries = array();
@@ -50,6 +51,8 @@ for($i = 0; $i < count($_ids); $i++){
 
 foreach($queries as $table => $ids){
 	//don't clean all locks! - is this really a needed statement???
-	$DB_WE->query('DELETE FROM ' . LOCK_TABLE . ' WHERE tbl="' . $DB_WE->escape(stripTblPrefix($table)) . '" AND ID IN (' . implode(', ', $ids) . ') AND sessionID="' . session_id() . '" AND UserID=' . we_base_request::_(we_base_request::INT, 'we_cmd', 0, 2));
+	if($ids){
+		$DB_WE->query('DELETE FROM ' . LOCK_TABLE . ' WHERE tbl="' . $DB_WE->escape(stripTblPrefix($table)) . '" AND ID IN (' . implode(', ', $ids) . ') AND sessionID="' . session_id() . '" AND UserID=' . we_base_request::_(we_base_request::INT, 'we_cmd', 0, 2));
+	}
 }
 ?>UNLOCKED
